Remove unused view
Use proper empty state view for favorites
